所有Integer 改成int ibatis映射文件改成map/int

+++ trunk/Rax/src/rax/dao/ArticleContentDao.java
@@ -6,21 +6,21 @@ import org.springframework.dao.DataAccessException;
 
 public interface ArticleContentDao {
 
-    public Long create(String newInstance) throws DataAccessException;
+    public int create(String newInstance) throws DataAccessException;
 
     public int update(String transientObject) throws DataAccessException;
 
     public int delete(String persistentObject) throws DataAccessException;
 
-    public int deleteByCategoryId(Long id) throws DataAccessException;
+    public int deleteByCategoryId(int id) throws DataAccessException;
 
-    public String read(Long id) throws DataAccessException;
+    public String read(int id) throws DataAccessException;
 
-    public int countByCategoryId(Long id) throws DataAccessException;
+    public int countByCategoryId(int id) throws DataAccessException;
 
-    public List<String> listByCategoryId(Long id, int index, int num)
+    public List<String> listByCategoryId(int id, int index, int num)
             throws DataAccessException;
 
-    public List<String> listAllByCategoryId(Long id) throws DataAccessException;
+    public List<String> listAllByCategoryId(int id) throws DataAccessException;
 
 }
